@@ -33,8 +33,8 @@
         "@polkadot/typegen": "^10.13.1",
         "@types/jest": "^29.5.8",
         "@types/prompts": "^2.4.8",
-        "@typescript-eslint/eslint-plugin": "^5.62.0",
-        "@typescript-eslint/parser": "^5.62.0",
+        "@typescript-eslint/eslint-plugin": "^8.6.0",
+        "@typescript-eslint/parser": "^8.6.0",
         "eslint": "^8.57.0",
         "eslint-config-prettier": "^9.1.0",
         "esno": "^4.7.0",
@@ -44,7 +44,7 @@
         "ts-jest": "^29.2.4",
         "ts-node": "^10.9.1",
         "tsconfig-paths": "^4.2.0",
-        "typescript": "^5.5.2"
+        "typescript": "^5.6.2"
     },
     "dependencies": {
         "@polkadot/api": "^10.13.1",
