@@ -40,31 +40,7 @@ module.exports = {
                 default: 'array',
             },
         ],
-        '@typescript-eslint/ban-types': [
-            'error',
-            {
-                types: {
-                    Object: {
-                        message: 'Avoid using the `Object` type. Did you mean `object`?',
-                    },
-                    Function: {
-                        message: 'Avoid using the `Function` type. Prefer a specific function type, like `() => void`.',
-                    },
-                    Boolean: {
-                        message: 'Avoid using the `Boolean` type. Did you mean `boolean`?',
-                    },
-                    Number: {
-                        message: 'Avoid using the `Number` type. Did you mean `number`?',
-                    },
-                    String: {
-                        message: 'Avoid using the `String` type. Did you mean `string`?',
-                    },
-                    Symbol: {
-                        message: 'Avoid using the `Symbol` type. Did you mean `symbol`?',
-                    },
-                },
-            },
-        ],
+        '@typescript-eslint/no-wrapper-object-types': 'error',
         '@typescript-eslint/consistent-type-assertions': 'error',
         '@typescript-eslint/dot-notation': 'error',
         '@typescript-eslint/explicit-function-return-type': 'off',
